move reused function
