+++ apps/crn-server/test/data-providers/transformers/users.test.ts
@@ -1,5 +1,6 @@
 import {
   OrcidWorkCMS,
+  parseAwardsCollection,
   parseOrcidWorkFromCMS,
 } from '../../../src/data-providers/transformers';
 
@@ -37,5 +38,59 @@ describe('parseOrcidWorkFromCMS', () => {
     const orcidWorks = getOrcidWorks();
     delete orcidWorks.title;
     expect(parseOrcidWorkFromCMS(orcidWorks).title).toBeUndefined();
+  });
+});
+
+describe('parseAwardsCollection', () => {
+  it('maps award entries to name, date and icon url', () => {
+    expect(
+      parseAwardsCollection({
+        awardsCollection: {
+          items: [
+            {
+              date: '2024-01-01',
+              awardType: {
+                name: 'Open Science Champion',
+                icon: { url: 'https://example.com/badge.png' },
+              },
+            },
+          ],
+        },
+      }),
+    ).toEqual([
+      {
+        name: 'Open Science Champion',
+        date: '2024-01-01',
+        iconUrl: 'https://example.com/badge.png',
+      },
+    ]);
+  });
+
+  it('returns an empty array when there is no awards collection', () => {
+    expect(parseAwardsCollection({})).toEqual([]);
+  });
+
+  it('skips awards missing an award type name or a date', () => {
+    expect(
+      parseAwardsCollection({
+        awardsCollection: {
+          items: [
+            null,
+            { date: '2024-01-01', awardType: null },
+            { date: null, awardType: { name: 'Open Science Champion' } },
+            {
+              date: '2024-02-01',
+              awardType: { name: 'Open Science Champion', icon: null },
+            },
+          ],
+        },
+      }),
+    ).toEqual([
+      {
+        name: 'Open Science Champion',
+        date: '2024-02-01',
+        iconUrl: undefined,
+      },
+    ]);
   });
 });
